#  Changing BM25Retriever _retrieve to use numpy methods 