# Description

Using BM25Retriever with a modest  amount of data(~40000 short nodes) was taking several seconds, which seemed odd.

Apparently, after computing the scores of the nodes, the sorting is done with python's sorted, which is slow compared to numpy's argsort and sort. The supporting library rank_bm25 has a get_top_n function that can be adapted to do the _retrieve's work using these functions.

Fixes # BM25Retrieve retrieval performance

## New Package?

No

## Version Bump?

Did I bump the version in the `pyproject.toml` file of the package I am updating? (Except for the `llama-index-core` package)

- [x] Yes
- [ ] No

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)

## How Has This Been Tested?

make test, this little benchmark and checking the computed scores match in my use case.

```
import random
import string
from llama_index.retrievers.bm25 import BM25Retriever
from llama_index.core.schema import TextNode


nodes = []
for i in range(100000):
    nodes.append(TextNode( text= ''.join(random.choices(string.ascii_lowercase, k=20))))

bm25retriever = BM25Retriever.from_defaults(nodes=nodes,
                                            similarity_top_k= 5)

bm25retriever.retrieve("foobar")
```

before
```
In [5]: %timeit bm25retriever.retrieve("foobar")
746 ms ± 39.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```
after
```
: %timeit bm25retriever.retrieve("foobar")
20.2 ms ± 536 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
````
## Suggested Checklist:

- [x] I have performed a self-review of my own code
- [x] My changes generate no new warnings
- [x] New and existing unit tests pass locally with my changes
- [x] I ran `make format; make lint` to appease the lint gods