@smonux 2024-06-10
That doesn't result in the same output :

In [5]: foo
Out[5]: array([0, 1, 2, 3, 4, 5, 6, 7, 8, 9])

In [6]: foo[-2:]
Out[6]: array([8, 9])

In [7]: foo[::-1][:2]
Out[7]: array([9, 8])

Also, the [::-1] doesn't seem to actually duplicate the array (not an
expert in numpy's vodoo):

In [8]: foo2 = foo[::-1]

In [9]: foo2.base
Out[9]: array([0, 1, 2, 3, 4, 5, 6, 7, 8, 9])

Regards,

El lun, 10 jun 2024 a las 18:11, Logan ***@***.***>)
escribió:

> ***@***.**** commented on this pull request.
> ------------------------------
>
> In
> llama-index-integrations/retrievers/llama-index-retrievers-bm25/llama_index/retrievers/bm25/base.py
> <https://github.com/run-llama/llama_index/pull/14015#discussion_r1633506428>
> :
>
> >      def _retrieve(self, query_bundle: QueryBundle) -> List[NodeWithScore]:
>          if query_bundle.custom_embedding_strs or query_bundle.embedding:
>              logger.warning("BM25Retriever does not support embeddings, skipping...")
>
> -        scored_nodes = self._get_scored_nodes(query_bundle.query_str)
> +        query = query_bundle.query_str
> +        tokenized_query = self._tokenizer(query)
> +        scores = self.bm25.get_scores(tokenized_query)
> +
> +        top_n = scores.argsort()[::-1][: self._similarity_top_k]
>
> seems like it would be even faster if we didn't reverse the list
>
> top_n = scores.argsort()[-self._similarity_top_k:]
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/run-llama/llama_index/pull/14015#pullrequestreview-2108224776>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AUPSSVLHYTXRDWFOPKE5O2DZGXF4XAVCNFSM6AAAAABI7XUNROVHI2DSMVQWIX3LMV43YUDVNRWFEZLROVSXG5CSMV3GSZLXHMZDCMBYGIZDINZXGY>
> .
> You are receiving this because you authored the thread.Message ID:
> ***@***.***>
>